% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_normalizar}
\alias{cie_normalizar}
\title{Normalizar códigos CIE-10 (deprecated)}
\usage{
cie_normalizar(
  codigos = lifecycle::deprecated(),
  buscar_db = lifecycle::deprecated()
)
}
\arguments{
\item{codigos}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Character vector de códigos.
Use \code{\link[=cie_norm]{cie_norm()}} con \code{codes} en código nuevo.}

\item{buscar_db}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Logical, buscar código en DB
(default TRUE). Use \code{\link[=cie_norm]{cie_norm()}} con \code{search_db} en código nuevo.}
}
\value{
Character vector con códigos normalizados
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Alias en español de \code{\link[=cie_norm]{cie_norm()}}. Se mantiene por compatibilidad
con código existente en CRAN. Usar \code{\link[=cie_norm]{cie_norm()}} en código nuevo.
}
\examples{
# Deprecated: usar cie_norm()
cie_normalizar("E110")
}
\seealso{
Other validacion:
\code{\link[=cie_expand]{cie_expand()}},
\code{\link[=cie_norm]{cie_norm()}},
\code{\link[=cie_validate_vector]{cie_validate_vector()}}
}
\concept{validacion}
\keyword{internal}
