% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-siglas.R
\name{cie_siglas}
\alias{cie_siglas}
\title{Listar siglas médicas (deprecated)}
\usage{
cie_siglas(categoria = NULL)
}
\arguments{
\item{categoria}{Character opcional, filtrar por categoría}
}
\value{
tibble con columnas: sigla, termino_busqueda, categoria
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=cie_short]{cie_short()}}.
}
\seealso{
Other busqueda:
\code{\link[=cie_short]{cie_short()}}
}
\concept{busqueda}
\keyword{internal}
