% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciecl-package.R
\docType{package}
\name{ciecl-package}
\alias{ciecl}
\alias{ciecl-package}
\title{ciecl: International Classification of Diseases 'ICD-10'/'ICD-11' for Chile}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for working with the International Classification of Diseases ('ICD-10' Chile official 'MINSAL'/'DEIS' v2018). Includes optimized 'SQL' search with 'SQLite', fuzzy matching of medical terms ('Jaro-Winkler'), Charlson and Elixhauser comorbidity calculation, 'WHO' 'ICD-11' 'API' integration, and hierarchical code validation. Data from Centro FIC Chile 'DEIS' \url{https://deis.minsal.cl/centrofic/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/ciecl/}
  \item \url{https://github.com/ropensci/ciecl}
  \item Report bugs at \url{https://github.com/ropensci/ciecl/issues}
}

}
\author{
\strong{Maintainer}: Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID}) (Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud Pública, U. de Chile)

Authors:
\itemize{
  \item Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID}) (Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud Pública, U. de Chile)
}

Other contributors:
\itemize{
  \item Francisca Jofre (Package logo design) [contributor]
  \item Maëlle Salmon (Maëlle reviewed the package (v. 0.9.8) for rOpenSci, see <https://github.com/ropensci/software-review/issues/765>) [reviewer]
  \item Yanina Bellini (Yanina reviewed the package (v. 0.9.8) for rOpenSci, see <https://github.com/ropensci/software-review/issues/765>) [reviewer]
}

}
\keyword{internal}
