% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-sql.R
\name{cie10_clear_cache}
\alias{cie10_clear_cache}
\title{Limpiar cache SQLite (forzar rebuild)}
\usage{
cie10_clear_cache()
}
\value{
No return value, called for side effects (deletes SQLite cache).
}
\description{
Limpiar cache SQLite (forzar rebuild)
}
\examples{
# Ver ubicacion del cache
tools::R_user_dir("ciecl", "data")

\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
cie10_clear_cache() # Elimina cie10.db local
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=cie10_sql]{cie10_sql()}}, \code{\link[=cie10_disconnect]{cie10_disconnect()}}

Other sql_backend:
\code{\link[=cie10_disconnect]{cie10_disconnect()}},
\code{\link[=cie10_sql]{cie10_sql()}}
}
\concept{sql_backend}
