% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-sql.R
\name{cie10_disconnect}
\alias{cie10_disconnect}
\title{Cerrar conexion pooled SQLite}
\usage{
cie10_disconnect()
}
\value{
No return value, called for side effects.
}
\description{
Cierra la conexion reutilizable al archivo SQLite.
Util para liberar el lock del archivo .db.
}
\examples{
\donttest{
cie10_disconnect()
}
}
