% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-sql.R
\name{cie10_sql}
\alias{cie10_sql}
\title{Ejecutar consultas SQL sobre CIE-10 Chile}
\usage{
cie10_sql(query, close = lifecycle::deprecated())
}
\arguments{
\item{query}{String SQL valido SQLite (SELECT/WHERE/JOIN)}

\item{close}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Ignorado - la conexion
es pooled y se gestiona automaticamente. Sera eliminado en una
version futura.}
}
\value{
tibble resultado query
}
\description{
Ejecutar consultas SQL sobre CIE-10 Chile
}
\examples{
# Buscar diabetes
cie10_sql("SELECT codigo, descripcion FROM cie10 WHERE codigo LIKE 'E11\%'")

\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Contar por capitulo
cie10_sql("SELECT capitulo, COUNT(*) n FROM cie10 GROUP BY capitulo")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=cie10_clear_cache]{cie10_clear_cache()}}, \code{\link[=cie10_disconnect]{cie10_disconnect()}},
\code{\link[=cie_search]{cie_search()}}

Other sql_backend:
\code{\link[=cie10_clear_cache]{cie10_clear_cache()}},
\code{\link[=cie10_disconnect]{cie10_disconnect()}}
}
\concept{sql_backend}
