% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_expand}
\alias{cie_expand}
\title{Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)}
\usage{
cie_expand(codigo)
}
\arguments{
\item{codigo}{String codigo padre (ej. "E11")}
}
\value{
Character vector con todos los codigos hijos del codigo padre.
Vector vacio si el codigo no existe en la base de datos.
}
\description{
Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)
}
\examples{
cie_expand("E11")
}
\seealso{
\code{\link{cie_normalizar}}, \code{\link{cie_lookup}}

Other validacion: 
\code{\link{cie_normalizar}()},
\code{\link{cie_validate_vector}()}
}
\concept{validacion}
