% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_expand}
\alias{cie_expand}
\title{Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)}
\usage{
cie_expand(codigo)
}
\arguments{
\item{codigo}{String codigo padre (ej. "E11")}
}
\value{
Character vector codigos hijos
}
\description{
Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)
}
\examples{
cie_expand("E11")
}
