% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_expand}
\alias{cie_expand}
\title{Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)}
\usage{
cie_expand(code, codigo = lifecycle::deprecated())
}
\arguments{
\item{code}{String codigo padre (ej. "E11")}

\item{codigo}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{code}.}
}
\value{
Character vector con todos los codigos hijos del codigo padre.
Vector vacio si el codigo no existe en la base de datos.
}
\description{
Expandir codigo jerarquico (ej. E11 -> E11.0-E11.9)
}
\examples{
cie_expand("E11")
}
\seealso{
\code{\link[=cie_norm]{cie_norm()}}, \code{\link[=cie_lookup]{cie_lookup()}}

Other validacion:
\code{\link[=cie_norm]{cie_norm()}},
\code{\link[=cie_normalizar]{cie_normalizar()}},
\code{\link[=cie_validate_vector]{cie_validate_vector()}}
}
\concept{validacion}
