% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-search.R
\name{cie_guia_busqueda}
\alias{cie_guia_busqueda}
\title{Guia de funciones de busqueda CIE-10}
\usage{
cie_guia_busqueda()
}
\value{
data.frame con guia de uso
}
\description{
Muestra tabla comparativa de cuando usar cada funcion de busqueda.
}
\examples{
cie_guia_busqueda()
}
