% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-guide.R
\name{cie_guia_busqueda}
\alias{cie_guia_busqueda}
\title{Guía de funciones de búsqueda (deprecated)}
\usage{
cie_guia_busqueda()
}
\value{
tibble con guía comparativa
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=cie_guide]{cie_guide()}}.
}
\examples{
# Deprecated: usar cie_guide()
cie_guia_busqueda()
}
\seealso{
Other search:
\code{\link[=cie_describe]{cie_describe()}},
\code{\link[=cie_guide]{cie_guide()}},
\code{\link[=cie_lookup]{cie_lookup()}},
\code{\link[=cie_search]{cie_search()}}
}
\concept{search}
\keyword{internal}
