% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-search.R
\name{cie_guia_busqueda}
\alias{cie_guia_busqueda}
\title{Guia de funciones de busqueda (deprecated)}
\usage{
cie_guia_busqueda()
}
\value{
tibble con guia comparativa
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=cie_guide]{cie_guide()}}.
}
\examples{
\dontrun{
# Deprecated: usar cie_guide()
cie_guia_busqueda()
}
}
\seealso{
Other busqueda: 
\code{\link{cie_describe}()},
\code{\link{cie_guide}()},
\code{\link{cie_lookup}()},
\code{\link{cie_search}()},
\code{\link{cie_short}()},
\code{\link{cie_siglas}()}
}
\concept{busqueda}
\keyword{internal}
