% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-guide.R
\name{cie_guide}
\alias{cie_guide}
\title{Guia de funciones de busqueda CIE-10}
\usage{
cie_guide()
}
\value{
tibble con guia comparativa de funciones de busqueda
}
\description{
Muestra tabla comparativa de cuando usar cada funcion de busqueda.
}
\examples{
cie_guide()
}
\seealso{
\code{\link[=cie_search]{cie_search()}}, \code{\link[=cie_lookup]{cie_lookup()}}, \code{\link[=cie_short]{cie_short()}}

Other search:
\code{\link[=cie_describe]{cie_describe()}},
\code{\link[=cie_guia_busqueda]{cie_guia_busqueda()}},
\code{\link[=cie_lookup]{cie_lookup()}},
\code{\link[=cie_search]{cie_search()}}
}
\concept{search}
