% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-comorbid.R
\name{cie_map_comorbid}
\alias{cie_map_comorbid}
\title{Mapeo manual grupos comorbilidad Chile-especifico}
\usage{
cie_map_comorbid(codigos)
}
\arguments{
\item{codigos}{Character vector codigos CIE-10}
}
\value{
tibble con codigo + categoria_comorbilidad
}
\description{
Agrupa codigos CIE-10 chilenos en categorias comorbilidad MINSAL.
Basado en Decreto 1301/2016 MINSAL + icd::icd10_map_charlson.
}
\examples{
cie_map_comorbid(c("E11.0", "I50.9", "C50.9"))
}
