% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-comorbid.R
\name{cie_map_comorbid}
\alias{cie_map_comorbid}
\title{Mapeo manual grupos comorbilidad Chile-especifico}
\usage{
cie_map_comorbid(codes, codigos = lifecycle::deprecated())
}
\arguments{
\item{codes}{Character vector de codigos}

\item{codigos}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{codes}.}
}
\value{
tibble con columnas: codigo, categoria
}
\description{
Agrupa codigos CIE-10 chilenos en categorias comorbilidad MINSAL.
Basado en Decreto 1301/2016 MINSAL + icd::icd10_map_charlson.
}
\examples{
cie_map_comorbid(c("E11.0", "I50.9", "C50.9"))
}
\seealso{
\code{\link{cie_comorbid}}, \code{\link{cie_norm}}

Other comorbilidades: 
\code{\link{cie_comorbid}()}
}
\concept{comorbilidades}
