% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_normalizar}
\alias{cie_normalizar}
\title{Normalizar codigos CIE-10 (deprecated)}
\usage{
cie_normalizar(codigos, buscar_db = TRUE)
}
\arguments{
\item{codigos}{Character vector de codigos}

\item{buscar_db}{Logical, buscar codigo en DB (default TRUE)}
}
\value{
Character vector con codigos normalizados
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Alias en espanol de \code{\link[=cie_normalize]{cie_normalize()}}. Se mantiene por compatibilidad
con codigo existente en CRAN. Usar \code{\link[=cie_normalize]{cie_normalize()}} en codigo nuevo.
}
\seealso{
Other validacion: 
\code{\link{cie_expand}()},
\code{\link{cie_normalize}()},
\code{\link{cie_validate_vector}()}
}
\concept{validacion}
\keyword{internal}
