% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-table.R
\name{cie_table}
\alias{cie_table}
\title{Generar tabla HTML interactiva GT de codigo CIE-10}
\usage{
cie_table(codigo, interactive = TRUE)
}
\arguments{
\item{codigo}{String codigo (ej. "E11" muestra jerarquia)}

\item{interactive}{Logical, tabla gt reactable (default TRUE)}
}
\value{
Objeto gt table HTML
}
\description{
Generar tabla HTML interactiva GT de codigo CIE-10
}
\examples{
\dontrun{
cie_table("E11")  # Diabetes tipo 2 completo
}
}
