% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie-utils.R
\name{cie_validate_vector}
\alias{cie_validate_vector}
\title{Validar vector de codigos CIE-10 formato}
\usage{
cie_validate_vector(codigos, strict = FALSE)
}
\arguments{
\item{codigos}{Character vector codigos (ej. c("E11.0", "Z00.0"))}

\item{strict}{Logical, validar existencia en DB (default FALSE)}
}
\value{
Logical vector misma longitud input
}
\description{
Validar vector de codigos CIE-10 formato
}
\examples{
cie_validate_vector(c("E11.0", "INVALIDO", "Z00"))
}
