% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciecl-package.R
\docType{package}
\name{ciecl-package}
\alias{ciecl}
\alias{ciecl-package}
\title{ciecl: International Classification of Diseases 'ICD-10'/'ICD-11' for Chile}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for working with the International Classification of Diseases ('ICD-10' Chile official 'MINSAL'/'DEIS' v2018). Includes optimized 'SQL' search with 'SQLite', fuzzy matching of medical terms ('Jaro-Winkler'), Charlson and Elixhauser comorbidity calculation, 'WHO' 'ICD-11' 'API' integration, and hierarchical code validation. Data from Centro FIC Chile 'DEIS' \url{https://deis.minsal.cl/centrofic/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rodotasso.github.io/ciecl/}
  \item \url{https://github.com/RodoTasso/ciecl}
  \item Report bugs at \url{https://github.com/RodoTasso/ciecl/issues}
}

}
\author{
\strong{Maintainer}: Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID})

}
\keyword{internal}
